Move `AddDarkMode` command
