+++ wcfsetup/install/files/lib/command/style/AddDarkMode.class.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace wcf\system\style\command;
+namespace wcf\command\style;
 
 use wcf\data\style\Style;
 use wcf\data\style\StyleEditor;
@@ -10,34 +10,27 @@ use wcf\system\style\StyleHandler;
 /**
  * Adds the dark color scheme to a style.
  *
- * @author Alexander Ebert
- * @copyright 2001-2023 WoltLab GmbH
- * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
- * @since 6.0
+ * @author      Alexander Ebert
+ * @copyright   2001-2023 WoltLab GmbH
+ * @license     GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
+ * @since       6.2
  */
 final class AddDarkMode
 {
-    private readonly int $styleID;
-
-    private readonly StyleHandler $styleHandler;
-
-    public function __construct(Style $style)
-    {
-        $this->styleID = $style->styleID;
-
-        $this->styleHandler = StyleHandler::getInstance();
-    }
+    public function __construct(
+        private readonly Style $style
+    ) {}
 
     public function __invoke(): void
     {
-        $styleEditor = new StyleEditor(new Style($this->styleID));
+        $styleEditor = new StyleEditor($this->style);
         $styleEditor->update([
             'hasDarkMode' => 1,
         ]);
 
-        $style = new Style($this->styleID);
+        $style = new Style($this->style->styleID);
 
         StyleCacheBuilder::getInstance()->reset();
-        $this->styleHandler->resetStylesheet($style);
+        StyleHandler::getInstance()->resetStylesheet($style);
     }
 }
